int(round)改成round
python3.5里round已经返回整数了

+++ src/spatial_index/common_utils.py
@@ -195,8 +195,8 @@ class Geohash:
         """
         lat, lon, lat_err, lon_err = self.decode_exactly(geohash)
         # Format to the number of decimals that are known
-        lats = "%.*f" % (max(1, int(round(-log10(lat_err)))) - 1, lat)
-        lons = "%.*f" % (max(1, int(round(-log10(lon_err)))) - 1, lon)
+        lats = "%.*f" % (max(1, round(-log10(lat_err))) - 1, lat)
+        lons = "%.*f" % (max(1, round(-log10(lon_err))) - 1, lon)
         if '.' in lats: lats = lats.rstrip('0')
         if '.' in lons: lons = lons.rstrip('0')
         return lons, lats
